Add Texture.fromCanvasImageSource
The CanvasRenderingContext2D.drawImage() method accepts
any object implementing the CanvasImageSource interface.
This libray uses Canvas.Texture to represent such objects.

Currently this library provides Texture.fromDomImage to create a Texture
from an HTMLImageElement. fromDomImage however enforces that
image.tagName == "IMG" meaning it cannot be used for any of the other
interfaces that browsers accept as an CanvasImageSource like:

* SVGImageElement
* HTMLVideoElement
* HTMLCanvasElement
* ImageBitmap
* OffscreenCanvas

This commit therefore introduces a new Texture.fromCanvasImageSource
function that essentially does the same as Texture.fromDomImage but
without enforcing a tagName. Since this makes the fromDomImage function
redundant this commit also deprecates it.